chore: post 관련 type 추출
